switching to hash from search in the URL
I'm not sure if this is a good thing to do or not
Due to the fact that Firefox translates %3D into = when in the window.location.hash function, I needed to use a different symbol besides "=" to separate statements in the URL. So, I switched to using ~ instead of equals.

It doesn't seem that other encoding/decoding methods would be better, since the fundamental problem is the need to disambiguate between "=" as part of the key and "=" as a separator.