Updating A->B formatting for easier readability
Note that the streamId itself is used for CSS class selections, and therefore has a restricted character set so I couldn't use ">". To documentation suggests that in some cases unusual characters can be escaped and still used with selection. But this didn't work when using d3 to manage the selection. Not sure if it would work otherwise.

Therefore, I am using "---" to concatenate in the code and when used for css selection and using the arrow character when the streamId is visible in the legend, detail box, and butter bar.